Create custom hook useSearch for searchbar on navigation and search page
